prompt added for delete dialog box (#19)
* prompt added for delete dialog box

* updated to confirmation so as to not require user input

+++ src/main/js/components/subcomponents/MenuItem.js
@@ -170,7 +170,11 @@ class ManagerMenuItem extends React.Component {
      * defined in App.js.
      */
     handleDelete() {
-        this.props.onDelete(this.props.menuItem, 'menuItems');
+
+        let confirmation = confirm("Are you sure you want to delete item?");
+        if (confirmation == true){
+            this.props.onDelete(this.props.menuItem, 'menuItems');
+        }
     }
 
     /**
